ddl: split tests to speed up tests
