+++ ddl/db_test.go
@@ -26,7 +26,6 @@ import (
 
 	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
-	"github.com/pingcap/parser/ast"
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/mysql"
 	tmysql "github.com/pingcap/parser/mysql"
@@ -35,7 +34,6 @@ import (
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/meta/autoid"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx"
@@ -123,123 +121,6 @@ func (s *testDBSuite) testErrorCode(c *C, sql string, errCode int) {
 	tErr, ok := originErr.(*terror.Error)
 	c.Assert(ok, IsTrue, Commentf("err: %T", originErr))
 	c.Assert(tErr.ToSQLError().Code, DeepEquals, uint16(errCode), Commentf("MySQL code:%v", tErr.ToSQLError()))
-}
-
-func (s *testDBSuite) TestMySQLErrorCode(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-
-	// create database
-	sql := "create database aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	s.testErrorCode(c, sql, tmysql.ErrTooLongIdent)
-	sql = "create database test"
-	s.testErrorCode(c, sql, tmysql.ErrDBCreateExists)
-	sql = "create database test1 character set uft8;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	sql = "create database test2 character set gkb;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	sql = "create database test3 character set laitn1;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	// drop database
-	sql = "drop database db_not_exist"
-	s.testErrorCode(c, sql, tmysql.ErrDBDropExists)
-	// create table
-	s.tk.MustExec("create table test_error_code_succ (c1 int, c2 int, c3 int, primary key(c3))")
-	sql = "create table test_error_code_succ (c1 int, c2 int, c3 int)"
-	s.testErrorCode(c, sql, tmysql.ErrTableExists)
-	sql = "create table test_error_code1 (c1 int, c2 int, c2 int)"
-	s.testErrorCode(c, sql, tmysql.ErrDupFieldName)
-	sql = "create table test_error_code1 (c1 int, aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa int)"
-	s.testErrorCode(c, sql, tmysql.ErrTooLongIdent)
-	sql = "create table aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa(a int)"
-	s.testErrorCode(c, sql, tmysql.ErrTooLongIdent)
-	sql = "create table test_error_code1 (c1 int, c2 int, key aa (c1, c2), key aa (c1))"
-	s.testErrorCode(c, sql, tmysql.ErrDupKeyName)
-	sql = "create table test_error_code1 (c1 int, c2 int, c3 int, key(c_not_exist))"
-	s.testErrorCode(c, sql, tmysql.ErrKeyColumnDoesNotExits)
-	sql = "create table test_error_code1 (c1 int, c2 int, c3 int, primary key(c_not_exist))"
-	s.testErrorCode(c, sql, tmysql.ErrKeyColumnDoesNotExits)
-	sql = "create table test_error_code1 (c1 int not null default '')"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidDefault)
-	sql = "CREATE TABLE `t` (`a` double DEFAULT 1.0 DEFAULT 2.0 DEFAULT now());"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidDefault)
-	sql = "CREATE TABLE `t` (`a` double DEFAULT now());"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidDefault)
-	sql = "create table t1(a int) character set uft8;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	sql = "create table t1(a int) character set gkb;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	sql = "create table t1(a int) character set laitn1;"
-	s.testErrorCode(c, sql, tmysql.ErrUnknownCharacterSet)
-	sql = "create table test_error_code (a int not null ,b int not null,c int not null, d int not null, foreign key (b, c) references product(id));"
-	s.testErrorCode(c, sql, tmysql.ErrWrongFkDef)
-	sql = "create table test_error_code_2;"
-	s.testErrorCode(c, sql, tmysql.ErrTableMustHaveColumns)
-	sql = "create table test_error_code_2 (unique(c1));"
-	s.testErrorCode(c, sql, tmysql.ErrTableMustHaveColumns)
-	sql = "create table test_error_code_2(c1 int, c2 int, c3 int, primary key(c1), primary key(c2));"
-	s.testErrorCode(c, sql, tmysql.ErrMultiplePriKey)
-	sql = "create table test_error_code_3(pt blob ,primary key (pt));"
-	s.testErrorCode(c, sql, tmysql.ErrBlobKeyWithoutLength)
-	sql = "create table test_error_code_3(a text, unique (a(3073)));"
-	s.testErrorCode(c, sql, tmysql.ErrTooLongKey)
-	sql = "create table test_error_code_3(`id` int, key `primary`(`id`));"
-	s.testErrorCode(c, sql, tmysql.ErrWrongNameForIndex)
-	sql = "create table t2(c1.c2 blob default null);"
-	s.testErrorCode(c, sql, tmysql.ErrWrongTableName)
-	sql = "create table t2 (id int default null primary key , age int);"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidDefault)
-	sql = "create table t2 (id int null primary key , age int);"
-	s.testErrorCode(c, sql, tmysql.ErrPrimaryCantHaveNull)
-	sql = "create table t2 (id int default null, age int, primary key(id));"
-	s.testErrorCode(c, sql, tmysql.ErrPrimaryCantHaveNull)
-	sql = "create table t2 (id int null, age int, primary key(id));"
-	s.testErrorCode(c, sql, tmysql.ErrPrimaryCantHaveNull)
-
-	sql = "create table t2 (id int primary key , age int);"
-	s.tk.MustExec(sql)
-
-	// add column
-	sql = "alter table test_error_code_succ add column c1 int"
-	s.testErrorCode(c, sql, tmysql.ErrDupFieldName)
-	sql = "alter table test_error_code_succ add column aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa int"
-	s.testErrorCode(c, sql, tmysql.ErrTooLongIdent)
-	sql = "alter table test_comment comment 'test comment'"
-	s.testErrorCode(c, sql, tmysql.ErrNoSuchTable)
-	sql = "alter table test_error_code_succ add column `a ` int ;"
-	s.testErrorCode(c, sql, tmysql.ErrWrongColumnName)
-	s.tk.MustExec("create table test_on_update (c1 int, c2 int);")
-	sql = "alter table test_on_update add column c3 int on update current_timestamp;"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidOnUpdate)
-	sql = "create table test_on_update_2(c int on update current_timestamp);"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidOnUpdate)
-
-	// drop column
-	sql = "alter table test_error_code_succ drop c_not_exist"
-	s.testErrorCode(c, sql, tmysql.ErrCantDropFieldOrKey)
-	s.tk.MustExec("create table test_drop_column (c1 int );")
-	sql = "alter table test_drop_column drop column c1;"
-	s.testErrorCode(c, sql, tmysql.ErrCantRemoveAllFields)
-	// add index
-	sql = "alter table test_error_code_succ add index idx (c_not_exist)"
-	s.testErrorCode(c, sql, tmysql.ErrKeyColumnDoesNotExits)
-	s.tk.Exec("alter table test_error_code_succ add index idx (c1)")
-	sql = "alter table test_error_code_succ add index idx (c1)"
-	s.testErrorCode(c, sql, tmysql.ErrDupKeyName)
-	// drop index
-	sql = "alter table test_error_code_succ drop index idx_not_exist"
-	s.testErrorCode(c, sql, tmysql.ErrCantDropFieldOrKey)
-	sql = "alter table test_error_code_succ drop column c3"
-	s.testErrorCode(c, sql, int(tmysql.ErrUnknown))
-	// modify column
-	sql = "alter table test_error_code_succ modify testx.test_error_code_succ.c1 bigint"
-	s.testErrorCode(c, sql, tmysql.ErrWrongDBName)
-	sql = "alter table test_error_code_succ modify t.c1 bigint"
-	s.testErrorCode(c, sql, tmysql.ErrWrongTableName)
-	// insert value
-	s.tk.MustExec("create table test_error_code_null(c1 char(100) not null);")
-	sql = "insert into test_error_code_null (c1) values(null);"
-	s.testErrorCode(c, sql, tmysql.ErrBadNull)
 }
 
 func (s *testDBSuite) TestAddIndexAfterAddColumn(c *C) {
@@ -1400,65 +1281,6 @@ func match(c *C, row []interface{}, expected ...interface{}) {
 	}
 }
 
-func (s *testDBSuite) TestUpdateMultipleTable(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("create database umt_db")
-	tk.MustExec("use umt_db")
-	tk.MustExec("create table t1 (c1 int, c2 int)")
-	tk.MustExec("insert t1 values (1, 1), (2, 2)")
-	tk.MustExec("create table t2 (c1 int, c2 int)")
-	tk.MustExec("insert t2 values (1, 3), (2, 5)")
-	ctx := tk.Se.(sessionctx.Context)
-	dom := domain.GetDomain(ctx)
-	is := dom.InfoSchema()
-	db, ok := is.SchemaByName(model.NewCIStr("umt_db"))
-	c.Assert(ok, IsTrue)
-	t1Tbl, err := is.TableByName(model.NewCIStr("umt_db"), model.NewCIStr("t1"))
-	c.Assert(err, IsNil)
-	t1Info := t1Tbl.Meta()
-
-	// Add a new column in write only state.
-	newColumn := &model.ColumnInfo{
-		ID:                 100,
-		Name:               model.NewCIStr("c3"),
-		Offset:             2,
-		DefaultValue:       9,
-		OriginDefaultValue: 9,
-		FieldType:          *types.NewFieldType(tmysql.TypeLonglong),
-		State:              model.StateWriteOnly,
-	}
-	t1Info.Columns = append(t1Info.Columns, newColumn)
-
-	kv.RunInNewTxn(s.store, false, func(txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		_, err = m.GenSchemaVersion()
-		c.Assert(err, IsNil)
-		c.Assert(m.UpdateTable(db.ID, t1Info), IsNil)
-		return nil
-	})
-	err = dom.Reload()
-	c.Assert(err, IsNil)
-
-	tk.MustExec("update t1, t2 set t1.c1 = 8, t2.c2 = 10 where t1.c2 = t2.c1")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("8 1", "8 2"))
-	tk.MustQuery("select * from t2").Check(testkit.Rows("1 10", "2 10"))
-
-	newColumn.State = model.StatePublic
-
-	kv.RunInNewTxn(s.store, false, func(txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		_, err = m.GenSchemaVersion()
-		c.Assert(err, IsNil)
-		c.Assert(m.UpdateTable(db.ID, t1Info), IsNil)
-		return nil
-	})
-	err = dom.Reload()
-	c.Assert(err, IsNil)
-
-	tk.MustQuery("select * from t1").Check(testkit.Rows("8 1 9", "8 2 9"))
-	tk.MustExec("drop database umt_db")
-}
-
 func (s *testDBSuite) TestCreateTableTooLarge(c *C) {
 	s.tk = testkit.NewTestKit(c, s.store)
 	s.tk.MustExec("use test")
@@ -1592,92 +1414,6 @@ func (s *testDBSuite) TestTableForeignKey(c *C) {
 	failSQL = "alter table t1 add foreign key (c) REFERENCES t3(a);"
 	s.testErrorCode(c, failSQL, tmysql.ErrKeyColumnDoesNotExits)
 	s.tk.MustExec("drop table if exists t1,t2,t3;")
-}
-
-func (s *testDBSuite) TestBitDefaultValue(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("create table t_bit (c1 bit(10) default 250, c2 int);")
-	tk.MustExec("insert into t_bit set c2=1;")
-	tk.MustQuery("select bin(c1),c2 from t_bit").Check(testkit.Rows("11111010 1"))
-	tk.MustExec("drop table t_bit")
-	tk.MustExec(`create table testalltypes1 (
-    field_1 bit default 1,
-    field_2 tinyint null default null
-	);`)
-	tk.MustExec(`create table testalltypes2 (
-    field_1 bit null default null,
-    field_2 tinyint null default null,
-    field_3 tinyint unsigned null default null,
-    field_4 bigint null default null,
-    field_5 bigint unsigned null default null,
-    field_6 mediumblob null default null,
-    field_7 longblob null default null,
-    field_8 blob null default null,
-    field_9 tinyblob null default null,
-    field_10 varbinary(255) null default null,
-    field_11 binary(255) null default null,
-    field_12 mediumtext null default null,
-    field_13 longtext null default null,
-    field_14 text null default null,
-    field_15 tinytext null default null,
-    field_16 char(255) null default null,
-    field_17 numeric null default null,
-    field_18 decimal null default null,
-    field_19 integer null default null,
-    field_20 integer unsigned null default null,
-    field_21 int null default null,
-    field_22 int unsigned null default null,
-    field_23 mediumint null default null,
-    field_24 mediumint unsigned null default null,
-    field_25 smallint null default null,
-    field_26 smallint unsigned null default null,
-    field_27 float null default null,
-    field_28 double null default null,
-    field_29 double precision null default null,
-    field_30 real null default null,
-    field_31 varchar(255) null default null,
-    field_32 date null default null,
-    field_33 time null default null,
-    field_34 datetime null default null,
-    field_35 timestamp null default null
-	);`)
-}
-
-func (s *testDBSuite) TestTableDDLWithFloatType(c *C) {
-	s.tk.MustExec("use test")
-	s.tk.MustExec("drop table if exists t")
-	s.testErrorCode(c, "create table t (a decimal(1, 2))", tmysql.ErrMBiggerThanD)
-	s.testErrorCode(c, "create table t (a float(1, 2))", tmysql.ErrMBiggerThanD)
-	s.testErrorCode(c, "create table t (a double(1, 2))", tmysql.ErrMBiggerThanD)
-	s.mustExec(c, "create table t (a double(1, 1))")
-	s.testErrorCode(c, "alter table t add column b decimal(1, 2)", tmysql.ErrMBiggerThanD)
-	// add multi columns now not support, so no case.
-	s.testErrorCode(c, "alter table t modify column a float(1, 4)", tmysql.ErrMBiggerThanD)
-	s.testErrorCode(c, "alter table t change column a aa float(1, 4)", tmysql.ErrMBiggerThanD)
-	s.mustExec(c, "drop table t")
-}
-
-func (s *testDBSuite) TestTableDDLWithTimeType(c *C) {
-	s.tk.MustExec("use test")
-	s.tk.MustExec("drop table if exists t")
-	s.testErrorCode(c, "create table t (a time(7))", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "create table t (a datetime(7))", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "create table t (a timestamp(7))", tmysql.ErrTooBigPrecision)
-	_, err := s.tk.Exec("create table t (a time(-1))")
-	c.Assert(err, NotNil)
-	s.mustExec(c, "create table t (a datetime)")
-	s.testErrorCode(c, "alter table t add column b time(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t add column b datetime(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t add column b timestamp(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t modify column a time(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t modify column a datetime(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t modify column a timestamp(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t change column a aa time(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t change column a aa datetime(7)", tmysql.ErrTooBigPrecision)
-	s.testErrorCode(c, "alter table t change column a aa timestamp(7)", tmysql.ErrTooBigPrecision)
-	s.mustExec(c, "alter table t change column a aa datetime(0)")
-	s.mustExec(c, "drop table t")
 }
 
 func (s *testDBSuite) TestTruncateTable(c *C) {
@@ -2022,31 +1758,6 @@ func (s *testDBSuite) TestRebaseAutoID(c *C) {
 	s.testErrorCode(c, "alter table tidb.test2 add column b int auto_increment key, auto_increment=10;", tmysql.ErrUnknown)
 }
 
-func (s *testDBSuite) TestZeroFillCreateTable(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test")
-	s.tk.MustExec("drop table if exists abc;")
-	s.tk.MustExec("create table abc(y year, z tinyint(10) zerofill, primary key(y));")
-	is := s.dom.InfoSchema()
-	tbl, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("abc"))
-	c.Assert(err, IsNil)
-	var yearCol, zCol *model.ColumnInfo
-	for _, col := range tbl.Meta().Columns {
-		if col.Name.String() == "y" {
-			yearCol = col
-		}
-		if col.Name.String() == "z" {
-			zCol = col
-		}
-	}
-	c.Assert(yearCol, NotNil)
-	c.Assert(yearCol.Tp, Equals, mysql.TypeYear)
-	c.Assert(mysql.HasUnsignedFlag(yearCol.Flag), IsTrue)
-
-	c.Assert(zCol, NotNil)
-	c.Assert(mysql.HasUnsignedFlag(zCol.Flag), IsTrue)
-}
-
 func (s *testDBSuite) TestCheckColumnDefaultValue(c *C) {
 	s.tk = testkit.NewTestKit(c, s.store)
 	s.tk.MustExec("use test;")
@@ -2278,133 +1989,6 @@ func (s *testDBSuite) TestGetTableEndHandle(c *C) {
 	maxID, emptyTable = s.getMaxTableRowID(testCtx)
 	result.Check(testkit.Rows(fmt.Sprintf("%v", maxID)))
 	c.Assert(emptyTable, IsFalse)
-}
-
-func (s *testDBSuite) TestMultiRegionGetTableEndHandle(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("drop database if exists test_get_endhandle")
-	tk.MustExec("create database test_get_endhandle")
-	tk.MustExec("use test_get_endhandle")
-
-	tk.MustExec("create table t(a bigint PRIMARY KEY, b int)")
-	for i := 0; i < 1000; i++ {
-		tk.MustExec(fmt.Sprintf("insert into t values(%v, %v)", i, i))
-	}
-
-	// Get table ID for split.
-	dom := domain.GetDomain(tk.Se)
-	is := dom.InfoSchema()
-	tbl, err := is.TableByName(model.NewCIStr("test_get_endhandle"), model.NewCIStr("t"))
-	c.Assert(err, IsNil)
-	tblID := tbl.Meta().ID
-
-	d := s.dom.DDL()
-	testCtx := newTestMaxTableRowIDContext(c, d, tbl)
-
-	// Split the table.
-	s.cluster.SplitTable(s.mvccStore, tblID, 100)
-
-	maxID, emptyTable := s.getMaxTableRowID(testCtx)
-	c.Assert(emptyTable, IsFalse)
-	c.Assert(maxID, Equals, int64(999))
-
-	tk.MustExec("insert into t values(10000, 1000)")
-	maxID, emptyTable = s.getMaxTableRowID(testCtx)
-	c.Assert(emptyTable, IsFalse)
-	c.Assert(maxID, Equals, int64(10000))
-
-	tk.MustExec("insert into t values(-1, 1000)")
-	maxID, emptyTable = s.getMaxTableRowID(testCtx)
-	c.Assert(emptyTable, IsFalse)
-	c.Assert(maxID, Equals, int64(10000))
-}
-
-func (s *testDBSuite) getHistoryDDLJob(id int64) (*model.Job, error) {
-	var job *model.Job
-
-	err := kv.RunInNewTxn(s.store, false, func(txn kv.Transaction) error {
-		t := meta.NewMeta(txn)
-		var err1 error
-		job, err1 = t.GetHistoryDDLJob(id)
-		return errors.Trace(err1)
-	})
-
-	return job, errors.Trace(err)
-}
-
-func (s *testDBSuite) TestBackwardCompatibility(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("create database if not exists test_backward_compatibility")
-	defer tk.MustExec("drop database test_backward_compatibility")
-	tk.MustExec("use test_backward_compatibility")
-	tk.MustExec("create table t(a int primary key, b int)")
-	for i := 0; i < 200; i++ {
-		tk.MustExec(fmt.Sprintf("insert into t values(%v, %v)", i, i))
-	}
-
-	// alter table t add index idx_b(b);
-	is := s.dom.InfoSchema()
-	schemaName := model.NewCIStr("test_backward_compatibility")
-	tableName := model.NewCIStr("t")
-	schema, ok := is.SchemaByName(schemaName)
-	c.Assert(ok, IsTrue)
-	tbl, err := is.TableByName(schemaName, tableName)
-	c.Assert(err, IsNil)
-
-	// Split the table.
-	s.cluster.SplitTable(s.mvccStore, tbl.Meta().ID, 100)
-
-	unique := false
-	indexName := model.NewCIStr("idx_b")
-	idxColName := &ast.IndexColName{
-		Column: &ast.ColumnName{
-			Schema: schemaName,
-			Table:  tableName,
-			Name:   model.NewCIStr("b"),
-		},
-		Length: types.UnspecifiedLength,
-	}
-	idxColNames := []*ast.IndexColName{idxColName}
-	var indexOption *ast.IndexOption
-	job := &model.Job{
-		SchemaID:   schema.ID,
-		TableID:    tbl.Meta().ID,
-		Type:       model.ActionAddIndex,
-		BinlogInfo: &model.HistoryInfo{},
-		Args:       []interface{}{unique, indexName, idxColNames, indexOption},
-	}
-	txn, err := s.store.Begin()
-	c.Assert(err, IsNil)
-	t := meta.NewMeta(txn)
-	job.ID, err = t.GenGlobalID()
-	c.Assert(err, IsNil)
-	job.Version = 1
-	job.StartTS = txn.StartTS()
-
-	// Simulate old TiDB init the add index job, old TiDB will not init the model.Job.ReorgMeta field,
-	// if we set job.SnapshotVer here, can simulate the behavior.
-	job.SnapshotVer = txn.StartTS()
-	err = t.EnQueueDDLJob(job)
-	c.Assert(err, IsNil)
-	err = txn.Commit(context.Background())
-	c.Assert(err, IsNil)
-	ticker := time.NewTicker(s.lease)
-	for range ticker.C {
-		historyJob, err := s.getHistoryDDLJob(job.ID)
-		c.Assert(err, IsNil)
-		if historyJob == nil {
-
-			continue
-		}
-		c.Assert(historyJob.Error, IsNil)
-
-		if historyJob.IsSynced() {
-			break
-		}
-	}
-
-	// finished add index
-	tk.MustExec("admin check index t idx_b")
 }
 
 func (s *testDBSuite) TestColumnModifyingDefinition(c *C) {
