+++ ddl/ddl_test.go
@@ -100,6 +100,7 @@ func (d *ddl) restartWorkers(ctx context.Context) {
 
 func TestT(t *testing.T) {
 	CustomVerboseFlag = true
+	CustomParallelSuiteFlag = true
 	logLevel := os.Getenv("log_level")
 	logutil.InitLogger(&logutil.LogConfig{
 		Level:  logLevel,
